internal/download: move download logic to this package
also disable http2 when downloading, may fix some issue

For #1563
